HW-3 BaseActivity, BaseFragment 추가 및 ResourceID 전달 형태로 수정 & onViewCreate메소드 사용(MainFragment 36~49줄) & view.recyclerview_mainfragment.run{} 형태로 반복 코드 제거(MainFragment 39~42줄) & Presenter 생성시 message 전달
