+++ machine-human/AndroidArchitectureStudy/app/src/main/java/org/study/kotlin/androidarchitecturestudy/base/BaseActivity.kt
@@ -1,0 +1,16 @@
+package org.study.kotlin.androidarchitecturestudy.base
+
+import android.os.Bundle
+import android.support.annotation.LayoutRes
+import android.support.v7.app.AppCompatActivity
+
+abstract class BaseActivity(
+    @LayoutRes
+    private val layoutId: Int
+) : AppCompatActivity() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(layoutId)
+    }
+}
