+++ machine-human/AndroidArchitectureStudy/app/src/main/java/org/study/kotlin/androidarchitecturestudy/base/BaseFragment.kt
@@ -1,0 +1,18 @@
+package org.study.kotlin.androidarchitecturestudy.base
+
+import android.os.Bundle
+import android.support.annotation.LayoutRes
+import android.support.v4.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+abstract class BaseFragment(
+    @LayoutRes
+    private val layoutId: Int
+) : Fragment() {
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        var view: View? = inflater.inflate(layoutId, container, false)
+        return view
+    }
+}
